[CollectionView] Refactor IItemsViewSource and ItemsSourceFactory
