+++ Xamarin.Forms.Core/IItemsViewSource.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Xamarin.Forms
+{
+	public interface IItemsViewSource : IDisposable
+	{
+		int Count { get; }
+		object this[int index] { get; }
+	}
+}
