+++ Xamarin.Forms.Platform.Android/CollectionView/ObservableItemsSource.cs
@@ -11,12 +11,12 @@ namespace Xamarin.Forms.Platform.Android
 		readonly IList _itemsSource;
 		bool _disposed;
 
-		public ObservableItemsSource(IList itemSource, RecyclerView.Adapter adapter)
+		public ObservableItemsSource(INotifyCollectionChanged itemSource, RecyclerView.Adapter adapter)
 		{
-			_itemsSource = itemSource;
+			_itemsSource = (IList)itemSource;
 			_adapter = adapter;
 
-			((INotifyCollectionChanged)itemSource).CollectionChanged += CollectionChanged;
+			itemSource.CollectionChanged += CollectionChanged;
 		}
 
 		public int Count => _itemsSource.Count;
