+++ Xamarin.Forms.Platform.Android/Xamarin.Forms.Platform.Android.csproj
@@ -69,7 +69,6 @@
     <Compile Include="CollectionView\CarouselViewRenderer.cs" />
     <Compile Include="CollectionView\CenterSnapHelper.cs" />
     <Compile Include="CollectionView\DataChangeObserver.cs" />
-    <Compile Include="CollectionView\EmptySource.cs" />
     <Compile Include="CollectionView\NongreedySnapHelper.cs" />
     <Compile Include="CollectionView\SingleSnapHelper.cs" />
     <Compile Include="CollectionView\EmptyViewAdapter.cs" />
@@ -77,9 +76,7 @@
     <Compile Include="CollectionView\ItemsViewAdapter.cs" />
     <Compile Include="CollectionView\EdgeSnapHelper.cs" />
     <Compile Include="CollectionView\EndSnapHelper.cs" />
-    <Compile Include="CollectionView\IItemsViewSource.cs" />
     <Compile Include="CollectionView\ItemContentView.cs" />
-    <Compile Include="CollectionView\ItemsSourceFactory.cs" />
     <Compile Include="CollectionView\ItemsViewRenderer.cs" />
     <Compile Include="CollectionView\ListSource.cs" />
     <Compile Include="CollectionView\ObservableItemsSource.cs" />
