--- Xamarin.Forms.Platform.iOS/CollectionView/EmptySource.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace Xamarin.Forms.Platform.iOS
-{
-	sealed class EmptySource : IItemsViewSource
-	{
-		public int Count => 0;
-
-		public object this[int index] => throw new IndexOutOfRangeException("IItemsViewSource is empty");
-
-		public void Dispose()
-		{
-		}
-	}
-}
