--- Xamarin.Forms.Platform.iOS/CollectionView/IItemsViewSource.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Xamarin.Forms.Platform.iOS
-{
-	internal interface IItemsViewSource : IDisposable
-	{
-		int Count { get; }
-		object this[int index] { get; }
-	}
-}
